pyproject.toml
[tool.poetry]
authors = [
   "Tyler Bikaun <tyler.bikaun@research.uwa.edu.au>"
]
description = ""
exclude = ["tests/"]
name = "mtbf_from_mwo"
version = "0.0.1"
license = "MIT"
readme = "README.md"

[tool.poetry.scripts]
mtbf_from_mwo = "run:main"

[tool.poetry.dependencies]
python = "^3.7"
tqdm= "4.52.0"
pandas = "0.25.3"
numpy ="1.18.1"
matplotlib="3.3.2"
scipy ="1.4.1"
reliability="0.5.1"
weibull="0.1.3"
statsmodels="0.11.1"
pyxlsb="1.0.8"
gensim = "3.8.3"
PyYAML = "^5.4.1"
xlrd = "1.2.0"
openpyxl = "^3.0.7"

[tool.poetry.extras]

[tool.poetry.dev-dependencies]
autoflake = "*"
black = { version = "19.10b0", python = "^3.8", markers = "platform_python_implementation == 'CPython'" }
flake8 = "*"
yapf = "*"
isort = "^5.6.4"
mypy = "*"
pylint = "*"
pytest = "*"
pytest-cov = "*"
poetry2conda = "^0.3.0"
pytest-mock = "^3.5.1"

[tool.yapf]
based_on_style = "google"
spaces_before_comment = 4
indent_width = 2
split_before_logical_operator = "true"
column_limit = 80

[tool.pylint.FORMAT]
output_format = "colorized"

[tool.black]
line_length = 80
target_version = ['py38']

[tool.poetry2conda]
name = "mtbf_from_mwo"

[tool.pytest.ini_options]
minversion = "6.0"
addopts = "--quiet --color=auto"
norecursedirs = ".git dist build .tox *.egg-info"
testpaths = [
    "tests",
]

[build-system]
build_backend = "poetry.masonry.api"
requires = ["poetry>=1.1.0"]
